pyproject.toml
[project]
name = "chatbot-model-context-protocol"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-groq>=1.1.2",
    "mcp-use>=1.6.0",
    "python-dotenv>=1.2.1",
]
